factorized a few test utilities
